[fix: role]: fix route update role

+++ server/controllers/role.controller.js
@@ -1,5 +1,6 @@
 const Role = require("../models/Role");
 const Func = require("../models/Func");
+const Account = require("../models/Account");
 
 const getAllRoles = async (req, res) => {
   try {
@@ -90,43 +91,52 @@ const createRole = async (req, res) => {
   }
 };
 
-const updateFunctionsForRoleById = async (req, res) => {
+const updateRoleById = async (req, res) => {
   try {
-    const { funcs } = req.body;
+    const { roleName, funcs } = req.body;
 
-    let checker = await Role.findByIdAndUpdate(req.params.id, { funcs });
-    if (!checker) {
+    // Check if role exists
+    let checker = await Role.findById(req.params.id);
+    if (!checker)
       return res.status(404).json({
         success: false,
         message: "Role not found",
       });
-    }
-    return res.status(201).json({
-      success: true,
-      message: "Functions for this role have been updated successfully",
-    });
-  } catch (error) {
-    console.log(error);
-    return res.status(500).json({
-      success: false,
-      message: "Internal server error",
-    });
-  }
-};
 
-const updateRoleNameById = async (req, res) => {
-  try {
-    const { roleName } = req.body;
-    let checker = await Role.findByIdAndUpdate(req.params.id, { roleName });
-    if (!checker) {
-      return res.status(404).json({
-        success: false,
-        message: "Role not found",
+    // Delete all connections of this role with funcs before updating
+    if (checker.funcs) {
+      checker.funcs.forEach(async (func) => {
+        const checkFunc = await Func.findById(func);
+        if (checkFunc) {
+          await Func.findByIdAndUpdate(func, {
+            roles: checkFunc.roles.filter((role) => role !== checker._id),
+          }).then((result) => result.save());
+        }
       });
     }
+
+    // Update role's info
+    const updateRole = await Role.findByIdAndUpdate(req.params.id, {
+      roleName,
+      funcs,
+    });
+    updateRole.save();
+
+    // Add connections of this role with funcs
+    if (funcs) {
+      funcs.forEach(async (func) => {
+        const updateFunc = await Func.findById(func);
+        if (updateFunc) {
+          await Func.findByIdAndUpdate(func, {
+            roles: [...updateFunc.roles, updateRole._id],
+          }).then((result) => result.save());
+        }
+      });
+    }
+
     return res.status(201).json({
       success: true,
-      message: "Role's name has been updated successfully",
+      message: "Update role successfully",
     });
   } catch (error) {
     console.log(error);
@@ -139,13 +149,39 @@ const updateRoleNameById = async (req, res) => {
 
 const deleteRoleById = async (req, res) => {
   try {
-    let checker = await Role.findByIdAndDelete(req.params.id);
+    // Check if there are still accounts of this role
+    let checker = await Account.findOne({ role: req.params.id });
+    if (checker) {
+      return res.status(406).json({
+        success: false,
+        message: "Can not delete because there are still accounts of this role",
+      });
+    }
+
+    // Delete role
+    checker = await Role.findByIdAndDelete(req.params.id);
     if (!checker) {
       return res.status(404).json({
         success: false,
         message: "Role not found",
       });
     }
+
+    // Delete role's id in funcs
+    if (checker.funcs) {
+      checker.funcs.forEach(async (func) => {
+        const findFunc = await Func.findById(func);
+        const listRoleUpdate = findFunc.roles.filter(
+          (role) => role !== checker._id
+        );
+        await Func.findByIdAndUpdate(
+          func,
+          { roles: listRoleUpdate },
+          { new: true }
+        ).then((result) => result.save());
+      });
+    }
+
     return res.status(204).json({
       success: true,
       message: "Role has been deleted successfully",
@@ -163,7 +199,6 @@ module.exports = {
   getRoleById,
   getAllRoles,
   createRole,
-  updateFunctionsForRoleById,
-  updateRoleNameById,
+  updateRoleById,
   deleteRoleById,
 };
